refactor: notation runner
