+++ NotationV0/src/lib/crypto.ts
@@ -1,7 +1,7 @@
 import * as crypto from 'crypto';
 import * as fs from 'fs';
 
-// compute te checksum of the file
+// compute the checksum of the file
 export async function computeChecksum(filePath: string): Promise<string> {
   return new Promise((resolve, reject) => {
     const hash = crypto.createHash('sha256');
